requirements.txt
python-snap7
mysql-connector-python
flask
matplotlib
pandas